cli: move all the ruleset action callbacks into ruleset.c
Move all the callbacks to perform actions on the ruleset object into
ruleset.c. Prefix the function with bfc_ instead of bf_.
